Move kanban_status CRUD operations to a separate file
